flask-multi-db-monorepo/requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
pymongo
dnspython
Flask-Cors
python-dotenv
requests
openai
Jinja2
Werkzeug
bcrypt
mcp
mssql-python
azure-identity
azure-storage-blob
azure-cosmos
setuptools

